requirements.txt
claude-agent-sdk>=0.0.10
python-frontmatter>=1.0.0
pyyaml>=6.0
